requirements.txt
streamlit
openai>=1.0.0
pypdf
python-docx
scikit-learn
pandas
numpy
python-dotenv
